+++ app/src/main/java/ru/netology/nmedia/activity/MainActivity.kt
@@ -1,0 +1,38 @@
+package ru.netology.nmedia.activity
+
+import android.os.Bundle
+import androidx.activity.viewModels
+import androidx.appcompat.app.AppCompatActivity
+import ru.netology.nmedia.R
+import ru.netology.nmedia.databinding.ActivityMainBinding
+import ru.netology.nmedia.repository.PostService
+import ru.netology.nmedia.viewmodel.PostViewModel
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        val binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        val viewModel: PostViewModel by viewModels()
+        viewModel.data.observe(this) { post ->
+            with(binding) {
+                author.text = post.author
+                published.text = post.published
+                content.text = post.content
+                likes.text = PostService.showValues(post.likes)
+                shares.text = PostService.showValues(post.shares)
+                views.text = PostService.showValues(post.views)
+                like.setImageResource(
+                    if (post.likedByMe) R.drawable.ic_baseline_favorite_24 else R.drawable.ic_baseline_favorite_border_24
+                )
+            }
+        }
+        binding.like.setOnClickListener {
+            viewModel.like()
+        }
+        binding.share.setOnClickListener {
+            viewModel.share()
+        }
+    }
+
+}
