+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -1,0 +1,12 @@
+package ru.netology.nmedia.dto
+
+data class Post(
+    val id: Long,
+    val author: String,
+    val content: String,
+    val published: String,
+    var likes: Long,
+    var shares: Long,
+    var views: Long,
+    var likedByMe: Boolean
+)
